requirements.txt
streamlit
pyrebase4
pandas
altair
joblib
scikit-learn
nltk
python-dotenv

# HuggingFace + dependencies
transformers
torch
sentencepiece
